// WMS.Frontend/src/app/features/employees/models/employee.model.ts
export interface Employee {
  employeeId: number;
  firstName: string;
  lastName: string;
  fullName: string;
  email: string;
  phoneNumber: string;
  gender: string;
  dob: string;
  doj: string;
  departmentId: number;
  departmentName: string;
  roleId: number;
  roleName: string;
  status: string;
}